refacto: rename _conf to conf
