+++ pypel/__init__.py
@@ -6,7 +6,7 @@ import pypel.extractors as extractors
 from pypel.main import process_from_config
 from pypel.utils.elk.init_index import *
 from pypel.utils.elk.clean_index import *
-from pypel._config.config import set_config, get_config
+from pypel.config.config import set_config, get_config
 
 __all__ = ["processes", "ProcessFactory", "process_from_config", "init_index", "clean_index", "extractors",
            "loaders", "logger", "set_config", "get_config"]
